refactor: - use i64 insted i32 as id - use join for media - improve error handling

+++ bowerbird_core/src/model/mod.rs
@@ -6,15 +6,15 @@ pub mod pixiv;
 
 #[derive(Clone, Default, Debug, Deserialize, Serialize, PartialEq, Eq, FromRow)]
 pub struct Tag {
-    pub id: i32,
+    pub id: i64,
     pub alias: Vec<String>,
 }
 
 #[derive(Clone, Default, Debug, Deserialize, Serialize, PartialEq, Eq, FromRow)]
 pub struct Item<E, H> {
-    pub id: i32,
+    pub id: i64,
     #[sqlx(default)]
-    pub parent_id: Option<i32>,
+    pub parent_id: Option<i64>,
     pub source_id: Option<String>,
     pub source_inaccessible: bool,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -25,7 +25,7 @@ pub struct Item<E, H> {
     pub updated_at: Option<DateTime<Utc>>,
 
     #[sqlx(default)]
-    pub tag_ids: Vec<i32>,
+    pub tag_ids: Vec<i64>,
 
     #[sqlx(flatten)]
     pub extension: E,
@@ -38,9 +38,9 @@ pub struct Item<E, H> {
 
 #[derive(Clone, Default, Debug, Deserialize, Serialize, PartialEq, Eq, FromRow)]
 pub struct History<H> {
-    pub history_id: i32,
+    pub history_id: i64,
     #[sqlx(default)]
-    pub item_id: Option<i32>,
+    pub item_id: Option<i64>,
     #[sqlx(default)]
     pub updated_at: Option<DateTime<Utc>>,
     #[sqlx(flatten)]
@@ -49,7 +49,7 @@ pub struct History<H> {
 
 #[derive(Clone, Default, Debug, Deserialize, Serialize, PartialEq, Eq, FromRow)]
 pub struct Media<E> {
-    pub id: i32,
+    pub id: i64,
     pub url: Option<String>,
     pub size: Option<i32>,
     pub mime: Option<String>,
